improve routing tests

+++ test/updateComuni.js
@@ -5,7 +5,7 @@ const should = require('chai').should(); // eslint-disable-line
 const updater = require('../src/comuniUpdater');
 
 
-module.exports = function () {
+module.exports = function() {
   this.timeout(10000);
 
   const dataPath = `${__dirname}/../data`;
@@ -13,38 +13,36 @@ module.exports = function () {
   const jsonFile = `${dataPath}/${COMUNI_JSON_FILE}`;
   const jsonFileBk = `${jsonFile}.bk`;
 
-  describe('when json data already exists', function () {
+  const jsonExists = fs.existsSync(jsonFile);
 
-    before(function () {
-      fs.copyFileSync(jsonFile, jsonFileBk);
+  if (jsonExists) {
+    describe('when json data already exists', function() {
+      before(function () {
+        fs.copyFileSync(jsonFile, jsonFileBk);
+      });
+
+      after(function () {
+        fs.renameSync(jsonFileBk, jsonFile);
+      });
+
+      it('update without errors', (done) => {
+        updater((err) => {
+          if (err) {
+            err.should.be.equal(true);
+          }
+          done();
+        });
+      });
     });
 
-    after(function () {
-      fs.renameSync(jsonFileBk, jsonFile);
-    });
-
-    it('update without errors', (done) => {
-      updater((err) => {
-        console.log('YAY1', err);
-        if (err) {
+  } else {
+    describe('when there is no json data', function () {
+      it('update without errors', (done) => {
+        updater((err) => {
           err.should.be.equal(true);
-        }
-        done();
+          done();
+        });
       });
     });
-  });
-
-
-  describe('when there is no json data', function () {
-
-    it('update without errors', (done) => {
-      updater((err) => {
-        console.log('YAY2', err);
-        err.should.be.equal(true);
-        done();
-      });
-    });
-
-  });
-
+  }
 };
